+++ androidApp/src/main/java/br/com/concrete/composekmmmoviesapp/androidApp/data/DatabaseModule.kt
@@ -12,4 +12,8 @@ val databaseModule = module {
     factory {
         get<AppDatabase>().favoriteMoviesDao()
     }
+
+    single {
+        MovieDbRepository(get(), get(), get(), get())
+    }
 }
