+++ androidApp/src/main/java/br/com/concrete/composekmmmoviesapp/androidApp/data/MoviesDbApi.kt
@@ -1,7 +1,7 @@
 package br.com.concrete.composekmmmoviesapp.androidApp.data
 
-import br.com.concrete.composekmmmoviesapp.androidApp.data.dto.GenreDto
-import br.com.concrete.composekmmmoviesapp.androidApp.data.dto.MoviePageDto
+import br.com.concrete.composekmmmoviesapp.androidApp.data.apidto.GenreDto
+import br.com.concrete.composekmmmoviesapp.androidApp.data.apidto.MoviePageDto
 import retrofit2.Response
 import retrofit2.http.GET
