+++ androidApp/src/main/java/br/com/concrete/composekmmmoviesapp/androidApp/data/apidto/GenreDto.kt
@@ -1,4 +1,4 @@
-package br.com.concrete.composekmmmoviesapp.androidApp.data.dto
+package br.com.concrete.composekmmmoviesapp.androidApp.data.apidto
 
 data class GenreDto(
     val genres: List<GenreItemDto>
