+++ androidApp/src/main/java/br/com/concrete/composekmmmoviesapp/androidApp/home/favoritestab/FavoritesViewModel.kt
@@ -1,15 +1,10 @@
 package br.com.concrete.composekmmmoviesapp.androidApp.home.favoritestab
 
 import androidx.lifecycle.ViewModel
-import androidx.lifecycle.map
-import br.com.concrete.composekmmmoviesapp.androidApp.data.db.FavoriteMovieDbDao
-import br.com.concrete.composekmmmoviesapp.androidApp.data.mapper.FavoritesMapper
+import br.com.concrete.composekmmmoviesapp.androidApp.data.MovieDbRepository
 
 class FavoritesViewModel(
-    favoritesDao: FavoriteMovieDbDao,
-    private val favoritesMapper: FavoritesMapper
+    movieDbRepository: MovieDbRepository
 ) : ViewModel() {
-    val favoritesLiveData = favoritesDao.getFavoriteMovies().map {
-        favoritesMapper.mapDbToMovies(it)
-    }
+    val favoritesLiveData = movieDbRepository.getFavoriteMovies()
 }
