+++ androidApp/src/main/java/br/com/concrete/composekmmmoviesapp/androidApp/home/moviestab/MoviesViewModel.kt
@@ -5,11 +5,7 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.liveData
 import androidx.lifecycle.viewModelScope
-import br.com.concrete.composekmmmoviesapp.androidApp.R
-import br.com.concrete.composekmmmoviesapp.androidApp.data.MoviesDbApi
-import br.com.concrete.composekmmmoviesapp.androidApp.data.db.FavoriteMovieDbDao
-import br.com.concrete.composekmmmoviesapp.androidApp.data.mapper.FavoritesMapper
-import br.com.concrete.composekmmmoviesapp.androidApp.data.mapper.MoviesMapper
+import br.com.concrete.composekmmmoviesapp.androidApp.data.MovieDbRepository
 import br.com.concrete.composekmmmoviesapp.androidApp.data.model.Movie
 import kotlinx.coroutines.launch
 
@@ -20,34 +16,28 @@ sealed class MoviesListUiState {
 }
 
 class MoviesViewModel(
-    private val api: MoviesDbApi,
-    private val mapper: MoviesMapper,
-    private val favoriteDao: FavoriteMovieDbDao,
-    private val favoritesMapper: FavoritesMapper
+    private val moviesDbRepository: MovieDbRepository
 ) : ViewModel() {
     val moviesList: LiveData<MoviesListUiState> = liveData {
         emit(MoviesListUiState.Loading)
-        try {
-            val popularMovies = api.getPopularMovies().body()
-            val genres = api.getGenres().body()
-            if (popularMovies != null && genres != null) {
-                emit(MoviesListUiState.Success(mapper.mapMoviesDtoToMovie(popularMovies, genres)))
-            }
-        } catch (ex: Throwable) {
-            emit(MoviesListUiState.Error(R.string.generic_error))
-        }
+        val resultPopularMovies = moviesDbRepository.getPopularMovies()
+
+//        if (resultPopularMovies.isSuccess) {
+            emit(MoviesListUiState.Success(resultPopularMovies))
+//        } else {
+//            emit(MoviesListUiState.Error(R.string.generic_error))
+//        }
     }
 
     fun addToFavorite(movie: Movie) {
         viewModelScope.launch {
-            val movieDbEntity = favoritesMapper.mapMovieToDbEntity(movie)
-            favoriteDao.insertFavorite(movieDbEntity)
+            moviesDbRepository.addToFavorite(movie)
         }
     }
+
     fun removeFrom(movie: Movie) {
         viewModelScope.launch {
-            val movieDbEntity = favoritesMapper.mapMovieToDbEntity(movie)
-            favoriteDao.deleteFavorite(movieDbEntity)
+            moviesDbRepository.removeFromFavorite(movie)
         }
     }
 }
